wip: Activate, test (and fix) on_disk caching for proc-macro expansions!
Also add manual test.
