+++ compiler/rustc_middle/src/arena.rs
@@ -116,7 +116,7 @@ macro_rules! arena_types {
             [decode] specialization_graph: rustc_middle::traits::specialization_graph::Graph,
             [] crate_inherent_impls: rustc_middle::ty::CrateInherentImpls,
             [] hir_owner_nodes: rustc_hir::OwnerNodes<'tcx>,
-            [] token_stream: rustc_ast::tokenstream::TokenStream,
+            [decode] token_stream: rustc_ast::tokenstream::TokenStream,
         ]);
     )
 }
